+++ QuickFIXn/Fields/DateTimeField.cs
@@ -5,7 +5,7 @@ namespace QuickFix.Fields;
 
 public class DateTimeField : FieldBase<DateTime>
 {
-    protected readonly TimeStampPrecision TimePrecision = TimeStampPrecision.Millisecond;
+    protected readonly TimePrecision TimePrecision = TimePrecision.Millisecond;
 
     public DateTimeField(int tag)
         : base(tag, new DateTime()) {}
@@ -14,9 +14,9 @@ public class DateTimeField : FieldBase<DateTime>
         : base(tag, dt) {}
 
     public DateTimeField(int tag, DateTime dt, bool showMilliseconds)
-        : this(tag, dt, showMilliseconds ? TimeStampPrecision.Millisecond : TimeStampPrecision.Second ) { }
+        : this(tag, dt, showMilliseconds ? TimePrecision.Millisecond : TimePrecision.Second ) { }
 
-    public DateTimeField(int tag, DateTime dt, TimeStampPrecision timeFormatPrecision)
+    public DateTimeField(int tag, DateTime dt, TimePrecision timeFormatPrecision)
         : base(tag, dt)
     {
         TimePrecision = timeFormatPrecision;