+++ UnitTests/Fields/Converters/TimeOnlyConverterTests.cs
@@ -1,6 +1,6 @@
 using System;
 using NUnit.Framework;
-using QuickFix;
+using QuickFix.Fields;
 using QuickFix.Fields.Converters;
 
 namespace UnitTests.Fields.Converters;
@@ -20,9 +20,9 @@ public class TimeOnlyConverterTests
     public void ConvertToStringTest()
     {
         TimeOnly tOnly = new TimeOnly(13, 30, 45, 999, 50);
-        Assert.That(TimeOnlyConverter.Convert(tOnly, TimeStampPrecision.Second), Is.EqualTo("13:30:45"));
-        Assert.That(TimeOnlyConverter.Convert(tOnly, TimeStampPrecision.Millisecond), Is.EqualTo("13:30:45.999"));
-        Assert.That(TimeOnlyConverter.Convert(tOnly, TimeStampPrecision.Microsecond), Is.EqualTo("13:30:45.999050"));
-        Assert.That(TimeOnlyConverter.Convert(tOnly, TimeStampPrecision.Nanosecond), Is.EqualTo("13:30:45.999050000"));
+        Assert.That(TimeOnlyConverter.Convert(tOnly, TimePrecision.Second), Is.EqualTo("13:30:45"));
+        Assert.That(TimeOnlyConverter.Convert(tOnly, TimePrecision.Millisecond), Is.EqualTo("13:30:45.999"));
+        Assert.That(TimeOnlyConverter.Convert(tOnly, TimePrecision.Microsecond), Is.EqualTo("13:30:45.999050"));
+        Assert.That(TimeOnlyConverter.Convert(tOnly, TimePrecision.Nanosecond), Is.EqualTo("13:30:45.999050000"));
     }
 }